backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "openai>=1.86.0",
    "openai-agents>=0.0.17",
    "pdfplumber>=0.11.7",
    "pymongo>=4.13.1",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "uvicorn>=0.34.3",
]
